Fix[MQB]: misc warnings

+++ src/groups/mqb/mqbc/mqbc_recoverymanager.cpp
@@ -1255,7 +1255,8 @@ void RecoveryManager::setLiveDataSource(mqbnet::ClusterNode* source,
     // PRECONDITIONS
     BSLS_ASSERT_SAFE(source);
     BSLS_ASSERT_SAFE(0 <= partitionId);
-    BSLS_ASSERT_SAFE(partitionId < d_recoveryContextVec.size());
+    BSLS_ASSERT_SAFE(partitionId <
+                     static_cast<int>(d_recoveryContextVec.size()));
 
     RecoveryContext& recoveryCtx = d_recoveryContextVec[partitionId];
